requirements.txt
anyio==4.12.0
asyncpg==0.31.0
certifi==2025.11.12
fake-useragent==2.2.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==25.0
pillow==12.0.0
PyMuPDF==1.26.6
pytesseract==0.3.13
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
